requirements.txt
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
google-api-core==1.3.0
google-auth==1.5.1
google-cloud-speech==0.35.0
google-cloud-texttospeech==0.2.0
googleapis-common-protos==1.5.3
grpcio==1.14.1
idna==2.7
nltk==3.3
numpy==1.15.0
protobuf==3.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytz==2018.5
requests==2.19.1
rsa==3.4.2
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
urllib3==1.23
